+++ packages/kokoas-client/src/pages/projEstimate/estimateDataGrid/EstimateDataGrid.tsx
@@ -1,10 +1,11 @@
 import 'react-data-grid/lib/styles.css';
-import DataGrid, { CellKeyDownArgs, CellKeyboardEvent, DataGridHandle } from 'react-data-grid';
-import { useMemo, useRef } from 'react';
+import DataGrid from 'react-data-grid';
+import { useMemo } from 'react';
 import { RowItem, getColumns } from './columns';
 import { EstimateDataGridContainer } from './EstimateDataGridContainer';
-import { useChangeRows } from '../hooks/useChangeRows';
+import { useChangeRows } from './useChangeRows';
 import { KItem } from '../schema';
+import { useDataGridKeyCellKeyDown } from './useDataGridKeyCellKeyDown';
 
 
 
@@ -13,8 +14,11 @@ export const EstimatesDataGrid = () => {
     handleRowChange,
     fields,
   } = useChangeRows();
-  const dataGridRef = useRef<DataGridHandle>(null);
-  
+  const columns = useMemo(() => getColumns(), []);
+  const {
+    handleCellKeyDown,
+    dataGridRef,
+  } = useDataGridKeyCellKeyDown(columns);
 
   const fieldsWithIndex =  useMemo(
     ()=>{
@@ -23,56 +27,6 @@ export const EstimatesDataGrid = () => {
     [fields],
   );
 
-  const columns = useMemo(() => getColumns(), []);
-
-  const fieldsLength = fields.length;
-
-  function handleCellKeyDown(args: CellKeyDownArgs<RowItem>, event: CellKeyboardEvent) {
-   
-    const { column, rowIdx } = args;
-    const { idx, editable } = column;
-    const { key } = event;
-
-    const {
-      selectCell,
-    } = dataGridRef.current || {};
-
-    const preventDefault = () => {
-      event.preventGridDefault();
-      event.preventDefault();
-    };
-    
-
-    if (!selectCell) return;
-    if (args.mode === 'EDIT') {
-      if (key === 'Enter') {
-        selectCell({ rowIdx: args.rowIdx, idx: idx + 1 });
-        preventDefault();
-      }
-      return;
-    }
-
-
-    if ((key === 'ArrowRight' || key === 'Enter') && idx === columns.length - 1) {
-      if (fieldsLength === 0) return;
-      if (rowIdx === -1) {
-        selectCell({ rowIdx: 0, idx: 0 });
-      } else {
-        if (rowIdx === fieldsLength - 1) return;
-        selectCell({ rowIdx: rowIdx + 1, idx: 0 });
-      }
-      preventDefault();
-    } else if (key === 'ArrowLeft' && idx === 0) {
-      if (rowIdx === -1) return;
-      selectCell({ rowIdx: rowIdx - 1, idx: columns.length - 1 });
-      preventDefault();
-    } else if (key === 'Enter' && !editable) {
-
-      selectCell({ rowIdx, idx: idx + 1 });
-      //preventDefault();
-    }
-  
-  }
 
   return (
     <EstimateDataGridContainer>
